refactor: decrease duplication
