+++ core/helpers/hardhat.ts
@@ -4,6 +4,10 @@ import { EthereumProvider } from 'hardhat/types';
 import { ethers } from 'ethers';
 import { formatToHex } from './format';
 import { pad32, concat, stripZeros } from './hex';
+import hre from 'hardhat';
+import { setupRpcUrlAndGetNetworks } from '../src/rpc';
+import { createWalletFromPrivateKey } from '../src/signer';
+import { NETWORK, REMOTE_RPC_URL } from '../helpers/constants';
 
 export const overwriteValueInSlot = async (
     network: string,
@@ -42,4 +46,36 @@ export const overwriteUintMapping = async (
 ) => {
     const slotAddress = generateMappingSlotAddress(mappingStartSlot, key);
     await overwriteUintValue(network, provider, contractName, slotAddress, newValue);
+};
+
+export const resetNetwork = async (blockNumber: number, rpcUrl: string = REMOTE_RPC_URL) => {
+    await hre.network.provider.request({
+        method: 'hardhat_reset',
+        params: [
+            {
+                forking: {
+                    jsonRpcUrl: rpcUrl,
+                    blockNumber: blockNumber,
+                },
+            },
+        ],
+    });
+};
+
+export const createWalletForRpc = async (signerPrivateKey: string, network: string = NETWORK) => {
+    const local_rpc_url = process.env.LOCAL_RPC_URL || 'http://localhost:8545';
+    await setupRpcUrlAndGetNetworks(local_rpc_url);
+    await createWalletFromPrivateKey(signerPrivateKey, network);
+    return hre.network.provider;
+};
+
+export const resetBlockchainFork = async function (
+    blockNumber: number,
+    signerPrivateKey: string,
+    network: string = NETWORK,
+    rpcUrl: string = REMOTE_RPC_URL
+) {
+    await resetNetwork(blockNumber, rpcUrl);
+    const provider = await createWalletForRpc(signerPrivateKey, network);
+    return provider;
 };
